GCP_demo/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-multipart==0.0.5
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
streaming-form-data==1.8.1
urllib3==1.26.3
python-dateutil==2.8.1
google-api-python-client==2.19.1
google-auth==2.0.0
google-auth-httplib2==0.1.0
google-cloud-pubsub==2.8.0
google-cloud-storage==1.42.1
google-api-python-client==2.19.1